repush files

+++ practice/9_solidity_projects/script/3_erc721/DeployNFTStore.s.sol
@@ -7,7 +7,7 @@ import {NFTStore} from "../../src/3_erc721/NFTStore.sol";
 import {TORATokenInNFT} from "../../src/3_erc721/TORATokenInNFT.sol";
 
 contract DeployNFTStore is Script {
-    function run(uint amount) external returns (NFTERC721, NFTStore, TORATokenInNFT) {
+    function run(uint256 amount) external returns (NFTERC721, NFTStore, TORATokenInNFT) {
         vm.startBroadcast();
         NFTERC721 NFTERC721Instance = new NFTERC721();
         TORATokenInNFT TORATokenInNFTInstance = new TORATokenInNFT(amount);
